+++ hphp/runtime/ext/std/ext_std_errorfunc.h
@@ -42,26 +42,26 @@ int64_t HHVM_FUNCTION(hphp_debug_backtrace_hash, int64_t options = 0);
 void HHVM_FUNCTION(debug_print_backtrace, int64_t options = 0,
                                           int64_t limit = 0);
 Array HHVM_FUNCTION(error_get_last);
-bool HHVM_FUNCTION(error_log, const String& message, int message_type = 0,
+bool HHVM_FUNCTION(error_log, const String& message, int64_t message_type = 0,
                               const Variant& destination = uninit_variant,
                               const Variant& extra_headers = uninit_variant);
 int64_t HHVM_FUNCTION(error_reporting, const Variant& level = uninit_variant);
 bool HHVM_FUNCTION(restore_error_handler);
 bool HHVM_FUNCTION(restore_exception_handler);
 Variant HHVM_FUNCTION(set_error_handler, const Variant& error_handler,
-                      int error_types = ((int)ErrorMode::PHP_ALL |
+                      int64_t error_types = ((int)ErrorMode::PHP_ALL |
                                          (int)ErrorMode::STRICT));
 Variant HHVM_FUNCTION(set_exception_handler, const Variant& exception_handler);
 void HHVM_FUNCTION(hphp_set_error_page, const String& page);
 void HHVM_FUNCTION(hphp_throw_fatal_error, const String& error_msg);
 void HHVM_FUNCTION(hphp_clear_unflushed);
 bool HHVM_FUNCTION(trigger_error, const String& error_msg,
-                                  int error_type = (int)ErrorMode::USER_NOTICE);
+                                  int64_t error_type = (int)ErrorMode::USER_NOTICE);
 bool HHVM_FUNCTION(trigger_sampled_error, const String& error_msg,
-                   int sample_rate,
-                   int error_type = (int)ErrorMode::USER_NOTICE);
+                   int64_t sample_rate,
+                   int64_t error_type = (int)ErrorMode::USER_NOTICE);
 bool HHVM_FUNCTION(user_error, const String& error_msg,
-                               int error_type = (int)ErrorMode::USER_NOTICE);
+                               int64_t error_type = (int)ErrorMode::USER_NOTICE);
 
 ArrayData* debug_backtrace_jit(int64_t options);
 String debug_string_backtrace(bool skip, bool ignore_args = false,
@@ -70,4 +70,3 @@ String stringify_backtrace(const Array& bt, bool ignore_args);
 
 ///////////////////////////////////////////////////////////////////////////////
 }
-
